test: Basic test setup
Make sure the tests can be ran easily.
This is the first of a list of change aiming at improving the test setup.
We will try to keep the changes simple and iterate rather than making a
giant pull request that would be painful to be reviewed and merged.

Note that the `rpcUrls.test.js` file was added back to tracking and
leverages `process.env`. Alternatively we could add dotenv as a dev
dependency and use (untracked) `.env` files for RPC URLs.

A lot of tests are still failing but at least some pass.
Fixing tests will be addressed in follow up pull requests.

Tested with both Hardhat and Anvil exporting `ETH_RPC` as:
```
export ETH_RPC=http://127.0.0.1:8545
```

In follow up pull requests we want to:
- fix or skip broken tests
- run tests from the CI